+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Missing modules AudioRecordManager, Speak, UPnPPortmap, and WorkDirectory.
+- Month bitmap utility functions.
 
 ## [0.1.0] - 2023-02-04